% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arcachon_lai}
\alias{arcachon_lai}
\title{arcachon_lai}
\format{A MODISTools tidy data frame}
\usage{
arcachon_lai
}
\description{
MODIS leaf area index (LAI) around the French town of Arcachon
derived from the MODIS MOD15A2H product (band Lai_500m).
}
\keyword{datasets}
