% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subset.R
\name{get_subset}
\alias{get_subset}
\title{Download all available dates}
\usage{
get_subset(product = NULL, band = NULL, lat = NULL, lon = NULL,
  start = "2000-01-01", end = format(Sys.time(), "\%Y-\%m-\%d"),
  km_lr = 0, km_ab = 0, site_id = NULL)
}
\arguments{
\item{product}{a valid MODIS product name}

\item{band}{band to download (default = NULL, all bands)}

\item{lat}{latitude in decimal degrees}

\item{lon}{longitude in decimal degrees}

\item{start}{start date}

\item{end}{end date}

\item{km_lr}{km left-right to sample}

\item{km_ab}{km above-below to sample}

\item{site_id}{site id (overides lat / lon)}
}
\value{
A nested list containing the downloaded data and a descriptive
header with meta-data.
}
\description{
Lists all available dates for a MODIS Land Products Subset product
at a particular location.
}
\examples{

\donttest{
# list all available MODIS Land Products Subsets products
bands <- list_dates(product = "MOD11A2", lat =)
print(bands)

}
}
\keyword{Land}
\keyword{MODIS}
\keyword{Products}
\keyword{Subsets,}
\keyword{meta-data}
\keyword{products,}
