% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_bands.R
\name{list_bands}
\alias{list_bands}
\title{Download all available bands}
\usage{
list_bands(product = NULL)
}
\arguments{
\item{product}{a valid MODIS product name}
}
\value{
A data frame of all available bands for a MODIS Land
Products Subsets products
}
\description{
Lists all available bands for a MODIS Land Products Subset product.
}
\examples{

\donttest{
# list all available MODIS Land Products Subsets products
bands <- list_bands(product = "MOD11A2")
print(bands)

}
}
\keyword{Land}
\keyword{MODIS}
\keyword{Products}
\keyword{Subsets,}
\keyword{meta-data}
\keyword{products,}
