% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_conversion.R
\name{ll_to_bb}
\alias{ll_to_bb}
\title{Convert lat / lon coordinates to sf bounding box polygon}
\usage{
ll_to_bb(lat, lon, cell_size, nrows, ncols)
}
\arguments{
\item{lat}{latitude generated by \code{\link[MODISTools]{sin_to_ll}}}

\item{lon}{longitude generated by \code{\link[MODISTools]{sin_to_ll}}}

\item{cell_size}{cell size provided by MODISTools}

\item{nrows}{cell size provided by MODISTools}

\item{ncols}{cell size provided by MODISTools}
}
\description{
Convert lat / lon coordinates to sf bounding box polygon
}
\examples{

\donttest{
# Download some test data
subset <- mt_subset(product = "MOD11A2",
                        lat = 40,
                        lon = -110,
                        band = "LST_Day_1km",
                        start = "2004-01-01",
                        end = "2004-03-31",
                        progress = FALSE)

# convert sinusoidal to lat / lon
lat_lon <- sin_to_ll(subset$xllcorner, subset$yllcorner)

# bind with the original dataframe
subset <- cbind(subset, lat_lon)

# convert to bounding box
bb <- apply(subset, 1, function(x){
  ll_to_bb(lon = x['longitude_ll'],
           lat = x['latitude_ll'],
           cell_size = x['cellsize'],
           nrows = x['nrows'],
           ncols = x['ncols'])
})

head(bb)
}
}
\seealso{
\code{\link[MODISTools]{sin_to_ll}}
}
\keyword{Land}
\keyword{MODIS}
\keyword{Products}
\keyword{Subsets,}
\keyword{meta-data}
\keyword{products,}
