% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_products.R
\name{mt_products}
\alias{mt_products}
\title{Download all available products}
\usage{
mt_products()
}
\value{
A data frame of all available MODIS Land Products Subsets products
}
\description{
Lists all available MODIS Land Products Subset products.
}
\examples{

\donttest{
# list all available MODIS Land Products Subsets products
products <- mt_products()
head(products)
}

}
\seealso{
\code{\link[MODISTools]{mt_bands}}
\code{\link[MODISTools]{mt_sites}} \code{\link[MODISTools]{mt_dates}}
}
\keyword{Land}
\keyword{MODIS}
\keyword{Products}
\keyword{Subsets,}
\keyword{meta-data}
\keyword{products,}
