% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_sites.R
\name{mt_sites}
\alias{mt_sites}
\title{Download all available fixed sites}
\usage{
mt_sites()
}
\value{
A data frame of all available MODIS Land Products Subsets
pre-processed sites
}
\description{
Lists all available MODIS Land Products Subset pre-processed sites
}
\examples{

\donttest{
# list all available MODIS Land Products Subsets products
sites <- mt_sites()
print(head(sites))
}
}
\seealso{
\code{\link[MODISTools]{mt_product}}
\code{\link[MODISTools]{mt_bands}} \code{\link[MODISTools]{mt_dates}}
}
\keyword{Land}
\keyword{MODIS}
\keyword{Products}
\keyword{Subsets,}
\keyword{meta-data}
\keyword{products,}
