% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_conversion.R
\name{sin_to_latlon}
\alias{sin_to_latlon}
\title{Convert sinusoidal coordinates to lat / lon}
\usage{
sin_to_latlon(x, y)
}
\arguments{
\item{x}{sinusoidal x coordinate (vector)}

\item{y}{sinusoidal y coordinate (vector)}
}
\description{
Convert sinusoidal coordinates to lat / lon
}
\examples{

\donttest{
# Download some test data
subset <- mt_subset(product = "MOD11A2",
                        lat = 40,
                        lon = -110,
                        band = "LST_Day_1km",
                        start = "2004-01-01",
                        end = "2004-03-31")

# convert sinusoidal to lat / lon
subset <- subset \%>\%
    mutate(result = map2(xllcorner, yllcorner, corner_to_coord)) \%>\%
    unnest()
}
}
\seealso{
[latlon_to_square()] [square_to_poly()]
}
\keyword{Land}
\keyword{MODIS}
\keyword{Products}
\keyword{Subsets,}
\keyword{meta-data}
\keyword{products,}
