% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_conversion.R
\name{square_to_poly}
\alias{square_to_poly}
\title{Convert square coordinates to sf bounding box polygon}
\usage{
square_to_poly(m)
}
\arguments{
\item{m}{matrix as generated by [latlon_to_square()]}
}
\description{
Convert square coordinates to sf bounding box polygon
}
\examples{

\donttest{
# Download some test data
subset <- mt_subset(product = "MOD11A2",
                        lat = 40,
                        lon = -110,
                        band = "LST_Day_1km",
                        start = "2004-01-01",
                        end = "2004-03-31")

# convert sinusoidal to lat / lon
subset <- subset \%>\%
    mutate(result = map2(xllcorner, yllcorner, corner_to_coord)) \%>\%
    unnest()


}
}
\seealso{
[latlon_to_square()] [sin_to_latlon()]
}
\keyword{Land}
\keyword{MODIS}
\keyword{Products}
\keyword{Subsets,}
\keyword{meta-data}
\keyword{products,}
