% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_subset.R
\name{write_subset}
\alias{write_subset}
\title{Write a MODISTools data structure to file}
\usage{
write_subset(df = NULL, out_dir = tempdir())
}
\arguments{
\item{df}{a nested data structure of class MODISTools}

\item{out_dir}{output directory where to store data}
}
\value{
writes MODISTools data structure to file, retains proper header info.
}
\description{
Writes a nested data structure of class MODISTools to file.
}
\examples{

\donttest{
# download data
subset = try(get_subset(product = "MOD11A2",
                        lat = 40,
                        lon = -110,
                        band = "LST_Day_1km",
                        start = "2004-01-01",
                        end = "2004-02-31"))

# print data structure
print(summary(df))

# write the phenocamo data file
write_subset(df, out_dir = tempdir())
}
}
\keyword{IO}
\keyword{series,}
\keyword{time}
