% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_products.R
\name{list_products}
\alias{list_products}
\title{Download all available products}
\usage{
list_products()
}
\value{
A data frame of all available MODIS Land Products Subsets products
}
\description{
Lists all available MODIS Land Products Subset products.
}
\examples{

\donttest{
# list all available MODIS Land Products Subsets products
products <- list_products()

#' # For other practical examples consult the included
# vignette.
}
}
\keyword{Land}
\keyword{MODIS}
\keyword{Products}
\keyword{Subsets,}
\keyword{meta-data}
\keyword{products,}
