% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_conversion.R
\name{ll_to_bb}
\alias{ll_to_bb}
\title{Convert lat / lon coordinates to sf bounding box polygon}
\usage{
ll_to_bb(lat, lon, cell_size, nrows, ncols)
}
\arguments{
\item{lat}{latitude generated by [sin_to_ll()]}

\item{lon}{longitude generated by [sin_to_ll()]}

\item{cell_size}{cell size provided by MODISTools}

\item{nrows}{cell size provided by MODISTools}

\item{ncols}{cell size provided by MODISTools}
}
\description{
Convert lat / lon coordinates to sf bounding box polygon
}
\examples{

\donttest{
# Download some test data
subset <- mt_subset(product = "MOD11A2",
                        lat = 40,
                        lon = -110,
                        band = "LST_Day_1km",
                        start = "2004-01-01",
                        end = "2004-03-31")

# convert sinusoidal to lat / lon
subset <- subset \%>\%
    mutate(result = map2(xllcorner, yllcorner, sin_to_ll)) \%>\%
    unnest()


}
}
\seealso{
[sin_to_ll()]
}
\keyword{Land}
\keyword{MODIS}
\keyword{Products}
\keyword{Subsets,}
\keyword{meta-data}
\keyword{products,}
