% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_sites.R
\name{mt_sites}
\alias{mt_sites}
\title{Download all available fixed sites}
\usage{
mt_sites(network)
}
\arguments{
\item{network}{the network for which to generate the site list,
when not provided the complete list is provided}
}
\value{
A data frame of all available MODIS Land Products Subsets
pre-processed sites
}
\description{
Lists all available MODIS Land Products Subset pre-processed sites
}
\examples{

\donttest{
# list all available MODIS Land Products Subsets products
sites <- mt_sites()
print(head(sites))
}

}
\seealso{
\code{\link[MODISTools]{mt_products}}
\code{\link[MODISTools]{mt_bands}} \code{\link[MODISTools]{mt_dates}}
}
